+++ CHANGELOG.md
@@ -14,13 +14,17 @@ v0.9.1.52 (February 16, 2021)
 
 **UI changes**
 
-* Add a button to system selected component page to remove a component from the system.
+* Added a button to system selected component page to remove a component from the system.
 * Moved "Add a component" to a system drop down to top of selected component page.
+* Added a search text box for each searching of components in the library and their statements.
+* Added pagnation to the component library and their statements.
+* Added a reset button for explicit resetting of component search.
 
 **Developer changes**
 
 * Move inclusion of `edit-component-modal.html` from `base.html` to `components/element_detail_tabs.html`.
 * Fix sort control order in `component_library_component` on the `components/element_detail_tabs.html` using the `natsort` package to sort SID correctly.
+* Comment out `controls.models.ElementControl.get_controls_by_element` method because it is not being used. Will delete after a few releases if not needed.
 
 **Data changes**